requirements.txt
docker-compose==1.8.0
